agent-service/gcp/disable-billing/requirements.txt
google-cloud-billing>=1.13.0
